feat(backend): Brickognize client and RecognizeService
Implements M1 Tasks 10-11:
- Async Brickognize client with retry and circuit breaker
- RecognizeService orchestrating rate limit, cache, color normalization, logging
- Recognize response schema

SQLite compatibility: ApiCallLog.id uses Integer autoincrement.

+++ backend/alembic/versions/0001_initial.py
@@ -28,7 +28,7 @@ def upgrade() -> None:
     )
     op.create_table(
         "api_call_log",
-        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
         sa.Column(
             "ts", sa.DateTime(timezone=True), nullable=False, server_default=sa.func.now()
         ),
